Refactor tests for core.add_reference
